+++ src/app/app.routes.ts
@@ -1,6 +1,5 @@
 import { Routes } from '@angular/router';
-import { BillDetailComponent } from '@bill/bill-detail/bill-detail.component';
-import { BillListComponent } from '@bill/bill-list/bill-list.component';
+import { BillDetailComponent, BillListComponent } from '@feature/bill';
 
 export const routes: Routes = [
 	{
