+++ src/app/feature/bill/index.ts
@@ -1,0 +1,6 @@
+export * from './bill-access.service';
+export * from './bill-create/bill-create.component';
+export * from './bill-detail/bill-detail.component';
+export * from './bill-list/bill-list.component';
+export * from './bill.model';
+export * from './bill.service';
