--- src/app/model/index.ts
@@ -1,3 +1,0 @@
-export * from './bill.model';
-export * from './expense.model';
-export * from './settlement.model';
